refactor: (#19) GameChat 컴포넌트 로직 분리(useGameChat) 및 UI 최적화
